Update to edit.ctp to fix related recipes not working
Had to add a forth parameter to the initAutocomplete() function so that we can pass in a regex. The regex is used to find the correct hidden input field to store the id of the ingredient or related recipe.  Before, it was just working for ingredients.  

Fixed the delete button for related recipes.  It had a similar problem, where it was only working for ingredients.